chore: fix unit test error
